Create index.html
This is the main page of Landscape Studios.  It promotes the program as well as gives the links to the various modules of the program, a description of each, and details about how to contribute.  If you have an idea for a new module, you may (1) comment here on GitHub, (2) tell me on the KhanAcademy equivalent of this page (https://www.khanacademy.org/computer-programming/landscape-studios/6689020578398208), or (3) just start working on it either by branching off on GitHub or spinning off of KhanAcademy.  If you make a spin off on KhanAcademy, but you also branch off on GitHub, then please start the name of the project with "GitHub:" (i.e. "GitHub: Landscape Studio" instead of "Spin-off of Landscape Studio" so that I know not to try to combine the thing your working on in KhanAcademy instead of through GitHub.